Stop dllexport instantiating ilist methods getPrevNode/getNextNode that trigger compile errors in headers
The parent type used for the template a incomplete in the headers the template
is use. If the header for the parent was included we would end up with circular
header dependency.
If a a class method has a default template parameter then it will not be implicitly
instantiated by any dllexport'ed classes deriving from the class template its
contained in.