[clang-tools-extra] Fix add_clang_library used without specify STATIC or SHARED
If a add_clang_library call doesn't specify building as static or shared library they are
implicitly added to the list static libraries that is linked in to clang-cpp shared library.
Because the clang-tools-extra libraries targets were declared after clang-cpp they
by luck never got linked to clang-cpp.
This change is required for clang symbol visibility macros on windows work
correctly for clang tools since we need to distinguish if a target being built
will be importing or exporting clang symbols from the clang-cpp DLL.